+++ tests/core/test_autodata.py
@@ -15,7 +15,7 @@ import pytest
 import torch
 from torch.utils.data import DataLoader, TensorDataset
 
-from gradsflow.core.autodata import AutoDataset
+from gradsflow.core.data import AutoDataset
 
 dataset = TensorDataset(torch.randn(8, 1, 32, 32))
 dataloader = DataLoader(dataset)
